Split overlap into primitive and contracted function
-> ContractedOrbital calls overlap_integral to normalize the linear combination. However, the same included file implemented overlap_integral for ContractedOrbital leading to conflict. This is now resolved by splitting primitive and contracted overlap.
