+++ include/Integrators/overlap_contracted.hpp
@@ -1,12 +1,8 @@
 #pragma once
 
-#include "BasisSet/slater_primitive.hpp"
-#include "BasisSet/gaussian_primitive.hpp"
+#include "Integrators/overlap_primitive.hpp"
 #include "BasisSet/contracted_orbital.hpp"
-
-double overlap_integral(const GaussianPrimitive& orbital1, const GaussianPrimitive& orbital2);
-
-double overlap_integral(const SlaterPrimitive& orbital1, const SlaterPrimitive& orbital2, const int n_offset = 0);
+#include "concepts.hpp"
 
 template <DerivedFromOrbital PrimitiveType>
 double overlap_integral(const ContractedOrbital<PrimitiveType>& orbital1, const ContractedOrbital<PrimitiveType>& orbital2) {
