+++ src/Integrators/overlap_primitive.cpp
@@ -2,7 +2,7 @@
 #include <cassert>
 #include <boost/math/special_functions/gamma.hpp>
 
-#include "Integrators/overlap_integral.hpp"
+#include "Integrators/overlap_primitive.hpp"
 #include "BasisSet/gaussian_primitive.hpp"
 #include "BasisSet/slater_primitive.hpp"
